releasetools: Store the build.prop file in the OTA zip
This file is often used to read information about the update contained
in the OTA. Place it in the update so it can be used by scripts. Add
build.prop to incrementals too

The file is not added to the updater-script, so it will not be placed
onto the actual system.

[GabrieleM: Adapt for A/B OTA packages]

Change-Id: Iabff12f559e62784c7b9012c5cb27514563b892f
(cherry picked from commit b987989c9007b37a9c28026345d4921d9b5f6a47)